app/services/users/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.4
flask-cors==3.0.6
flask-migrate==2.2.0
flask_bcrypt==0.7.1
pyjwt==1.6.4